requirements.txt
aiosmtplib==4.0.2
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
babel==2.17.0
backrefs==5.9
bcrypt==4.0.1
blinker==1.9.0
boto3==1.40.64
botocore==1.40.64
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fakeredis==2.32.0
fastapi==0.119.1
fastapi-cli==0.0.12
fastapi-cloud-cli==0.2.0
fastapi-mail==1.5.7
flake8==7.3.0
fonttools==4.60.1
ghp-import==2.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==7.0.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.22
mkdocs-material-extensions==1.3.1
numpy==2.3.3
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
pymdown-extensions==10.16.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
#python-magic-bin==0.4.14
python-multipart==0.0.20
PyYAML==6.0.2
pyyaml_env_tag==1.1
redis==6.4.0
regex==2025.10.23
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
s3transfer==0.14.0
sentry-sdk==2.38.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.48.0
typer==0.18.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
